// src/types/RUser.tsx
interface RUser {
  email: string;
  password: string;
  role: string;
  vegan: boolean;
  vegetarain: boolean;
  pescatarian: boolean;
  kosher: boolean;
  celiac_desease: boolean;
  peanut_allergy: boolean;
  dairy_allergy: boolean;
  seafood_allergy: boolean;
  egg_allergy: boolean;
}

export default RUser;